requirements.txt
pandas
numpy
scikit-learn
Flask
pytest
pyarrow
gunicorn
google-cloud-bigquery
google-cloud-storage
overrides